services/dream-service/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-cloud-pubsub==2.18.4
google-cloud-secret-manager==2.17.0
asyncpg==0.29.0
redis[hiredis]==5.0.1
python-multipart==0.0.6
python-json-logger==2.0.7
